structuring the Greenhouse Effect page code for the better
