+++ src/pages/efecto-invernadero/awareness-section/AwarenessKeyboardListeners.jsx
