+++ src/pages/efecto-invernadero/awareness-section/AwarenessText.jsx
