+++ src/pages/efecto-invernadero/awareness-section/animations/BombermanFireModel.jsx
