+++ src/pages/efecto-invernadero/awareness-section/animations/BurningAnimation.jsx
