+++ src/pages/efecto-invernadero/awareness-section/animations/BurningWorldModel.jsx
