+++ src/pages/efecto-invernadero/awareness-section/animations/ColdingAnimation.jsx
