+++ src/pages/efecto-invernadero/awareness-section/animations/DroughtAnimation.jsx
