+++ src/pages/efecto-invernadero/awareness-section/animations/DryMudModel.jsx
