+++ src/pages/efecto-invernadero/awareness-section/animations/FireModel.jsx
