+++ src/pages/efecto-invernadero/awareness-section/animations/PurpleRainstormModel.jsx
