+++ src/pages/efecto-invernadero/awareness-section/animations/RainAnimation.jsx
