+++ src/pages/efecto-invernadero/awareness-section/animations/SnowBallRocks.jsx
