+++ src/pages/efecto-invernadero/awareness-section/animations/SnowLight.jsx
