+++ src/pages/efecto-invernadero/solutions-section/BouncingBall.jsx
