+++ src/pages/efecto-invernadero/solutions-section/DodecahedronContainer.jsx
