+++ src/pages/efecto-invernadero/solutions-section/SolutionsAnimations.jsx
@@ -3,20 +3,20 @@ import { Physics } from "@react-three/rapier";
 
 import BouncingBall from "./BouncingBall";
 import DodecahedronContainer from "./DodecahedronContainer";
-import NormalGasAnimation from "./NormalGasAnimation";
+import NormalGasAnimation from "../awareness-section/animations/NormalGasAnimation";
 import SolutionsStage from "./SolutionsStage";
 import useGreeenhouseStore from "../../../stores/greenhouse-store";
 
-import Solution0 from "./Solution0";
-import Solution1 from "./Solution1";
-import Solution2 from "./Solution2";
-import Solution3 from "./Solution3";
-import Solution4 from "./Solution4";
-import Solution5 from "./Solution5";
-import Solution6 from "./Solution6";
-import Solution7 from "./Solution7";
-import Solution8 from "./Solution8";
-import Solution9 from "./Solution9";
+import Solution0 from "./solutions/Solution0";
+import Solution1 from "./solutions/Solution1";
+import Solution2 from "./solutions/Solution2";
+import Solution3 from "./solutions/Solution3";
+import Solution4 from "./solutions/Solution4";
+import Solution5 from "./solutions/Solution5";
+import Solution6 from "./solutions/Solution6";
+import Solution7 from "./solutions/Solution7";
+import Solution8 from "./solutions/Solution8";
+import Solution9 from "./solutions/Solution9";
 
 /**
  * SolutionsAnimations component
