+++ src/pages/efecto-invernadero/solutions-section/SolutionsStage.jsx
