+++ src/pages/efecto-invernadero/solutions-section/SolutionsText.jsx
