+++ src/pages/efecto-invernadero/solutions-section/solutions/BucketModel.jsx
