+++ src/pages/efecto-invernadero/solutions-section/solutions/PowerPlugModel.jsx
