+++ src/pages/efecto-invernadero/solutions-section/solutions/SolarPanelModel.jsx
