+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution0.jsx
