+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution1.jsx
