+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution2.jsx
