+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution6.jsx
