+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution7.jsx
