+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution8.jsx
