+++ src/pages/efecto-invernadero/solutions-section/solutions/Solution9.jsx
@@ -1,5 +1,3 @@
-import { RigidBody } from "@react-three/rapier";
-
 import { SocialMediaIcons } from "./SocialMediaIcons";
 
 /**
