+++ src/pages/efecto-invernadero/solutions-section/solutions/SpiralBulbModel.jsx
